+++ app/build.gradle
@@ -60,8 +60,6 @@ android {
     }
     buildFeatures {
         viewBinding true
-        //noinspection DataBindingWithoutKapt
-        dataBinding true
     }
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_11
@@ -118,15 +116,15 @@ dependencies {
     implementation 'org.nanohttpd:nanohttpd:2.3.1'
     implementation 'com.google.guava:guava:30.1.1-jre'
     // coroutines
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.1'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4'
     implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation 'io.github.g00fy2:versioncompare:1.5.0'
 
     testImplementation 'junit:junit:4.13.2'
     testImplementation 'com.google.truth:truth:1.1.3'
     testImplementation 'org.junit.jupiter:junit-jupiter:5.7.1'
-    testImplementation 'io.mockk:mockk:1.12.0'
+    testImplementation 'io.mockk:mockk:1.12.5'
     testImplementation 'org.robolectric:robolectric:4.4'
     testImplementation 'androidx.test:core:1.4.0'
     testImplementation "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
