+++ app/src/main/AndroidManifest.xml
@@ -94,10 +94,7 @@
         </activity-alias>
         <activity
             android:name=".ScanActivity"
-            android:theme="@style/OriginalTheme.Transparent" />
-        <activity
-            android:name=".PrivacyPolicyActivity"
-            android:theme="@style/OriginalTheme.Transparent" />
+            android:theme="@style/OriginalTheme.Dialog.Alert" />
 
         <service android:name=".PebbleService" />
         <service android:name=".BluetoothLeService" />
