+++ app/src/main/java/com/cooper/wheellog/MainActivity.java
@@ -37,7 +37,6 @@ import androidx.viewpager2.widget.ViewPager2;
 
 import com.cooper.wheellog.companion.WearOs;
 
-import com.cooper.wheellog.data.TripDao;
 import com.cooper.wheellog.data.TripDatabase;
 import com.cooper.wheellog.utils.Constants;
 import com.cooper.wheellog.utils.Constants.ALARM_TYPE;
@@ -88,7 +87,6 @@ public class MainActivity extends AppCompatActivity {
 
     protected static final int RESULT_DEVICE_SCAN_REQUEST = 20;
     protected static final int RESULT_REQUEST_ENABLE_BT = 30;
-    protected static final int ResultPrivatePolicy = 666;
 
     private static Boolean onDestroyProcess = false;
 
@@ -230,6 +228,7 @@ public class MainActivity extends AppCompatActivity {
                             if (WheelLog.AppConfig.getAutoWatch() && wearOs == null) {
                                 toggleWatch();
                             }
+                            DialogHelper.INSTANCE.checkPWMIsSetAndShowAlert(context);
                             WheelLog.Notifications.setNotificationMessageId(R.string.connected);
                             break;
                         case BluetoothLeService.STATE_DISCONNECTED:
@@ -495,9 +494,7 @@ public class MainActivity extends AppCompatActivity {
         setSupportActionBar(toolbar);
         audioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
 
-        if (!WheelLog.AppConfig.getPrivatePolicyAccepted()) {
-            startActivityForResult(new Intent(MainActivity.this, PrivacyPolicyActivity.class), ResultPrivatePolicy);
-        }
+        DialogHelper.INSTANCE.checkAndShowPrivatePolicyDialog(this);
 
         if (!getPackageManager().hasSystemFeature(PackageManager.FEATURE_BLUETOOTH_LE)) {
             Toast.makeText(this, R.string.ble_not_supported, Toast.LENGTH_SHORT).show();
@@ -519,11 +516,9 @@ public class MainActivity extends AppCompatActivity {
         registerReceiver(mCoreBroadcastReceiver, makeCoreIntentFilter());
         WheelLog.Notifications.update();
 
-        if (WheelLog.AppConfig.getDetectBatteryOptimization()) {
-            SomeUtil.Companion.checkBatteryOptimizationsAndShowAlert(this);
-        }
-
-        SomeUtil.Companion.checkPWMIsSetAndShowAlert(this);
+        DialogHelper.INSTANCE.checkBatteryOptimizationsAndShowAlert(this);
+        // for test without wheel go to isHardwarePWM and comment Unknown case
+        // DialogHelper.INSTANCE.checkPWMIsSetAndShowAlert(this);
     }
 
     @Override
@@ -864,13 +859,6 @@ public class MainActivity extends AppCompatActivity {
                     startBluetoothService();
                 else {
                     Toast.makeText(this, R.string.bluetooth_required, Toast.LENGTH_LONG).show();
-                    finish();
-                }
-                break;
-            case ResultPrivatePolicy:
-                if (resultCode == RESULT_OK) {
-                    WheelLog.AppConfig.setPrivatePolicyAccepted(true);
-                } else {
                     finish();
                 }
                 break;
