+++ app/src/main/java/com/cooper/wheellog/data/TripDao.kt
@@ -1,27 +1,28 @@
 package com.cooper.wheellog.data
 
+import androidx.lifecycle.LiveData
 import androidx.room.*
 
 @Dao
 interface TripDao {
     @Query("SELECT * FROM trip_database ORDER BY id ASC")
-    suspend fun getAll(): List<TripData>
+    fun getAll(): LiveData<List<TripData>>
 
     @Query("SELECT * FROM trip_database WHERE id IN (:tripIds) ORDER BY id ASC")
-    suspend fun loadAllByIds(tripIds: IntArray): List<TripData>
+    fun loadAllByIds(tripIds: IntArray): LiveData<List<TripData>>
 
     @Query("SELECT * FROM trip_database WHERE fileName LIKE :fileName LIMIT 1")
-    suspend fun getTripByFileName(fileName: String): TripData?
+    fun getTripByFileName(fileName: String): LiveData<TripData?>
 
     @Query("SELECT * FROM trip_database WHERE ecId LIKE :ecId LIMIT 1")
-    suspend fun getTripByElectroClubId(ecId: Int): TripData?
+    fun getTripByElectroClubId(ecId: Int): LiveData<TripData?>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insert(vararg tripData: TripData)
+    fun insert(vararg tripData: TripData)
 
     @Update
-    suspend fun update(vararg tripData: TripData)
+    fun update(vararg tripData: TripData)
 
     @Delete
-    suspend fun delete(vararg tripData: TripData)
+    fun delete(vararg tripData: TripData)
 }
