+++ app/src/main/res/layout/activity_scan.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    android:padding="5dp">
+    android:layout_margin="16dp">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -22,8 +22,7 @@
             android:ellipsize="end"
             android:singleLine="true"
             android:text="@string/scanning"
-            android:textAlignment="viewStart"
-            android:textColor="@color/wheelview_textblocks" />
+            android:textAlignment="viewStart" />
         <ProgressBar
             android:id="@+id/scanProgress"
             android:layout_width="wrap_content"
@@ -31,51 +30,24 @@
             android:gravity="end"
             android:visibility="gone"/>
     </LinearLayout>
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="0.5dp"
-        android:background="@color/divider" />
-
+    <View style="@style/Divider" />
     <ListView android:id="@android:id/list"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="0.5dp"
-        android:background="@color/divider" />
-
-    <LinearLayout
-        android:id="@+id/last_mac_layout"
+    <View style="@style/Divider" />
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/last_mac_text"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/wheelview_arc_dim"
-        android:orientation="horizontal"
-        android:padding="5dp">
+        app:endIconDrawable="@drawable/ic_baseline_login_24"
+        app:endIconTint="@color/primary_dark"
+        app:endIconCheckable="true"
+        app:endIconMode="custom">
 
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/last_mac_text"
-            android:layout_width="0dp"
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textColorHint="@color/wheelview_text">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/current_mac"
-                android:textColor="@color/wheelview_textblocks"
-                android:textColorHint="@color/wheelview_text" />
-        </com.google.android.material.textfield.TextInputLayout>
-
-        <Button
-            android:id="@+id/last_mac_ok"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:background="@android:color/transparent"
-            android:foreground="@drawable/ic_baseline_login_24"
-            android:textColorHint="@color/wheelview_text"
-            tools:ignore="UnusedAttribute" />
-    </LinearLayout>
+            android:hint="@string/current_mac" />
+    </com.google.android.material.textfield.TextInputLayout>
 </LinearLayout>
