+++ app/src/main/res/layout/scan_list_item.xml
@@ -5,9 +5,11 @@
     <TextView android:id="@+id/device_name"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:textColor="@color/primary_dark"
         android:textSize="24sp"/>
     <TextView android:id="@+id/device_address"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:textColor="@color/primary_dark"
         android:textSize="12sp"/>
 </LinearLayout>
